This commit has: Add Swimmers, Add Events, Sign up Swimmers to Events, and a Delete All Data link.
I've added accessibility to the Signup UX by requiring only the names to sign up for an event. The program retrieves the Swimmer IDs from the Swimmer table and associates them with the event in Signup.

Tomorrow I will be working on scores (times) by updating the "swim_time" column in the Signup table, and accessing and scoring each event.