Added all of my sample data. Added the race results through PSequel rather than cycling through my form for adding race results because the data I input wasn't being accepted. The PSequel table would show a new line had been created, and time stamp it, but add no data. When all I was adding was swim entry data it worked, but not results. I still think there is a conflict with handing times. Should have committed several changes ago.
Also tried to add a form so that the admin (or whoever) could enter the race id number and see the results for that race. I feel like I'm doing something wrong with the path. The results page works like a charm if I manually add a race id number to the url, but my form does not work.