Working on adding conference affiliations to college data. College table just has a conference ID, adding the conference name. The only data in the conference table is the conference_id and the conference name.
Also added links from the main page to each race status page (which includes the race result if the race has been completed) instead of having admin enter the race number in a form, which was throwing an error message. There was enough real estate on the page for the links, and it solves the functionality problem of being able to see race results. And one less step for the admin.