NoMethodError at /college_view undefined method `conference_name' for #College::ActiveRecord_Relation:0x007fcc8ea85158 file: delegation.rb location: method_missing line: 136
This is for an .erb page I am making just to display information about the participating colleges in the swim meet, including the conference affiliation. MyApp.get "/college_view" do is in my college_controller.rb file, and the erb page it is set up to send back is "college_status" . My colleges table does not have conference names, it has a conference_id foreign key that goes to the conference table. I need to use that to get the conference names. Conference_name is a method I added to the college.rb model page. Conference_name is intended to use the conference_id to access the conference table and return the conference name. Clearly I am doing something wrong. To me this is not that different from the presentation video of getting the director of a movie.