Worked with Sumeet on controller action for displaying top three finishers in an event (or showing a message that the event is not over yet) but I'm not sure if it's working properly.
Still need to add links to update/edit events, competitors, colleges, etc. (all forms for this are working).

Something is coded wrong in view all event details and awards. Whatever sequence number an event detail or award appears on the page, that's what Ruby thinks is the event_id or award_id (event though there will be multiple "100 meter freestyle" events in a row -- each is considered the next in the sequence and when you click on it it brings up the next event.) Need to fix this.

Need to start working on deletion that removes multiple pieces of info from database.